+++ statechanges/templates/statechanges/base.html
@@ -9,6 +9,7 @@
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
         integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
     <link rel="stylesheet" href="{% static 'css/statechange.css' %}">
+    <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
 </head>
 
 
@@ -18,17 +19,36 @@
             <img src="{% static 'images/navimage.png' %}" height="30" class="d-inline-block align-top" alt="">
             GET Protocol Fansite
         </a>
-            <ul class="navbar-nav">
-                <li class="nav-item {% if navbar == 'page_home' %}active{% endif %}">
-                    <a class="nav-link" href="{% url 'page_home' %}">Home <span class="sr-only">(current)</span></a>
-                </li>
-                <li class="nav-item {% if navbar == 'page_statechanges' %}active{% endif %}">
-                    <a class="nav-link" href="{% url 'page_statechanges' %}">Statechanges</a>
-                </li>
-                <li class="nav-item {% if navbar == 'page_events' %}active{% endif %}">
-                    <a class="nav-link" href="{% url 'page_events' %}">Events</a>
+        <div class="d-none d-lg-block">
+            <ul class="nav navbar-nav navbar-right">
+                    <li class="nav-item {% if navbar == 'page_home' %}active{% endif %}">
+                        <a class="nav-link" href="{% url 'page_home' %}">Home</a>
+                    </li>
+                    <li class="nav-item {% if navbar == 'page_statechanges' %}active{% endif %}">
+                        <a class="nav-link" href="{% url 'page_statechanges' %}">Statechanges</a>
+                    </li>
+                    <li class="nav-item {% if navbar == 'page_events' %}active{% endif %}">
+                        <a class="nav-link" href="{% url 'page_events' %}">Events</a>
+                    </li>
+            </ul>
+        </div>
+        <div class="d-lg-none">
+            <ul class="nav navbar-nav navbar-right">
+                <li class="nav-item dropdown">
+                    <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button"
+                        data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                        Menu
+                    </a>
+                    <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+                        <a class="dropdown-item" href="{% url 'page_home' %}">Home</a>
+                        <div class="dropdown-divider"></div>
+                        <a class="dropdown-item" href="{% url 'page_statechanges' %}">Statechanges</a>
+                        <div class="dropdown-divider"></div>
+                        <a class="dropdown-item" href="{% url 'page_events' %}">Events</a>
+                    </div>
                 </li>
             </ul>
+        </div>
     </nav>
     <div class="container">
