+++ statechanges/templates/statechanges/events.html
@@ -21,26 +21,28 @@
 
 <div class="row">
     <div class="col">
-        <div class="jumbotron">
-            <h1 class="display-7">Eventlist</h1>
-            <p class="lead">A list with all the events </p>
-            <p>Below you can find a list with all events using GET. The ones which have statechanges recently are on top.</p>
-        </div>
-    </div>
-</div>
-
-<div class="row">
-    <div class="col">
         <div id="accordion" role="tablist" aria-multiselectable="true">
             <div class="d-none d-sm-block">
                 <div class="card">
                     <h6 class="card-header" role="tab" id="headingOne">
                         <a data-parent="#accordion" class="d-block">
                             <div class="row no-gutters align-items-center">
-                                <div class="col-9">
+                                <div class="col-12">
+                                    <h1 class="display-7">All events</h1>
+                                    <p class="lead">Events with the most recent statechanges are on top.</p>
+                                </div>
+                            </div>
+                        </a>
+                    </h6>
+                </div>
+                <div class="card">
+                    <h6 class="card-header" role="tab" id="headingOne">
+                        <a data-parent="#accordion" class="d-block">
+                            <div class="row no-gutters align-items-center">
+                                <div class="col-10">
                                     <p><b>Event Details</b></p>
                                 </div>
-                                <div class="col-3">
+                                <div class="col-2">
                                     <p><b>Total Statechanges</b></p>
                                 </div>
                             </div>
@@ -49,19 +51,21 @@
                     </h6>
                 </div>
             </div>
+
             {% for pageevent in pageevents%}
             <div class="card">
                 <h8 class="card-header" role="tab" id="headingOne">
                     <a data-toggle="collapse" data-parent="#accordion" href="#batch{{ pageevent.hash }}"
                         aria-expanded="true" aria-controls="batch{{ pageevent.hash }}" class="d-block">
                         <div class="row no-gutters align-items-center">
+                            <div class="col-1">
+                                <p><i class="fa fa-calendar"></i></p>
+                            </div>
                             <div class="col-9">
-
                                 <p><b>Creation date:</b> {{ pageevent.block.date |date:"d-m-Y"}}</p>
                                 <p><b>Hash:</b> {{ pageevent.hash }}</p>
-
                             </div>
-                            <div class="col-3">
+                            <div class="col-2">
                                 <div class="d-sm-none">
                                     <p><b>State Changes:</b> {{ pageevent.totalsum }}</p>
                                 </div>
